ChatGui feedback
ChatGui or ChatGUI? Which is right? Hmm... - no one I know has come up with a (to me) satisfactory answer for whether to capitalize acronyms or leave them the same way (with camelCase) as the rest. I tend to avoid acronyms in my names for this reason.
Good use of scoping - the only variable you have set as a class field is the JFrame itself, and you've used local scope for everything that could be done. That helps keep it easier to keep track of what variables there are and when they can be used/what they're for.
Note that you could just extend JFrame from the class (public class ChatGui extends JFrame) and then you would not need a JFrame stored in the class.

The use of mouseClicked for your handler means that any kind of mouse click (left, right, middle) count the same on that button. It also means that (were tabbing to fields supported in your interface) you wouldn't be able to use the enter or space key to trigger the submit. A possibly more appropriate handler for this would be ActionEvent instead of MouseAdapter, since the default action for a button includes mouseClick, spacebar and enter simultaneously. No points lost, as this is functional as written.

Fields are split (remind me to talk about this) on whether to use variable scope parameters when not require or not - this means protected/private may not be required when working in a class in Java - my preference and the requirement for assignments will be to include them. No points lost this time, but you will lose points on it in later assignments.

One thing to note (which I will be bringing up in class) is that adding anything to the "getContentPane" of a JFrame is not widely recommended because it has a lot of problems it can cause. Usually you'd just do JFrame.add